+++ Assets/OkapiKit/Scripts/Editor/CombatTextSpawnPointEditor.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEditor;
+
+namespace OkapiKit.Editor
+{
+    [CustomEditor(typeof(CombatTextPivot))]
+    public class CombatTextSpawnPointEditor : OkapiBaseEditor
+    {
+        SerializedProperty propOffsetZ;
+
+        protected override void OnEnable()
+        {
+            base.OnEnable();
+        }
+
+        public override void OnInspectorGUI()
+        {
+            serializedObject.Update();
+
+            if (WriteTitle())
+            {
+                EditorGUI.BeginChangeCheck();
+
+                EditorGUI.EndChangeCheck();
+
+                serializedObject.ApplyModifiedProperties();
+                (target as OkapiElement).UpdateExplanation();
+            }
+        }
+
+        protected override GUIStyle GetTitleSyle()
+        {
+            return GUIUtils.GetActionTitleStyle();
+        }
+
+        protected override GUIStyle GetExplanationStyle()
+        {
+            return GUIUtils.GetActionExplanationStyle();
+        }
+
+        protected override string GetTitle()
+        {
+            return "Combat Text Pivot";
+        }
+
+        protected override Texture2D GetIcon()
+        {
+            return GUIUtils.GetTexture("CombatText");
+        }
+
+        protected override (Color, Color, Color) GetColors() => (GUIUtils.ColorFromHex("#d1d283"), GUIUtils.ColorFromHex("#2f4858"), GUIUtils.ColorFromHex("#969721"));
+    }
+}